코멘트 사항 반영, 리팩터링, 비즈니스로직 분리
