+++ src/main/java/subway/controller/dto/SectionResponse.java
@@ -1,0 +1,36 @@
+package subway.controller.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import subway.domain.Section;
+
+import static subway.controller.dto.StationResponse.stationToStationResponse;
+
+@Getter
+public class SectionResponse {
+    private Long id;
+    private Long lineId;
+    private StationResponse upStations;
+    private StationResponse downStations;
+    private Long distance;
+
+    @Builder
+    public SectionResponse(Long id, Long lineId, StationResponse upStations, StationResponse downStations, Long distance) {
+        this.id = id;
+        this.lineId = lineId;
+        this.upStations = upStations;
+        this.downStations = downStations;
+        this.distance = distance;
+    }
+
+    public static SectionResponse sectionToSectionResponse(Section section) {
+        return SectionResponse.builder()
+                .id(section.getId())
+                .lineId(section.getLine().getId())
+                .upStations(stationToStationResponse(section.getUpStation()))
+                .downStations(stationToStationResponse(section.getDownStation()))
+                .distance(section.getDistance())
+                .build();
+    }
+
+}
