+++ src/main/java/subway/controller/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package subway.station.request;
+package subway.controller.dto;
 
 public class StationRequest {
     private String name;
