+++ src/main/java/subway/domain/repository/SectionRepository.java
@@ -1,7 +1,7 @@
-package subway.section.repository;
+package subway.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import subway.section.entity.Section;
+import subway.domain.Section;
 
 public interface SectionRepository extends JpaRepository<Section, Long> {
 }
