--- src/main/java/subway/line/request/LineCreateRequest.java
@@ -1,24 +1,0 @@
-package subway.line.request;
-
-import lombok.Getter;
-import lombok.NonNull;
-
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
-
-@Getter
-public class LineCreateRequest {
-    @Size(max = 20)
-    @NotEmpty
-    private String name;
-    @Size(max = 20)
-    @NotEmpty
-    private String color;
-    @NotNull
-    private Long upStationId;
-    @NotNull
-    private Long downStationId;
-    @NotNull
-    private Long distance;
-}
