+++ src/test/java/subway/station/StationAcceptanceTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
-import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.jdbc.Sql;
 import subway.fixture.StationFixture;
 import subway.util.RestAssuredUtil;
@@ -14,8 +13,6 @@ import subway.util.RestAssuredUtil;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
-//@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @Sql("/truncate_table.sql")
 @DisplayName("지하철역 관련 기능")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
